requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asttokens==3.0.0
attrs==24.2.0
cachetools==5.5.0
cattrs==22.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
executing==2.1.0
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
highrise-bot-sdk==24.1.0
httplib2==0.22.0
icecream==2.1.3
idna==3.10
multidict==6.1.0
pendulum==2.1.2
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pyshout==0.2.3
python-dateutil==2.9.0.post0
pytube==15.0.0
pytzdata==2020.1
quattro==22.2.0
requests==2.32.3
rsa==4.9
six==1.16.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.17.1
youtube-dl==2021.12.17
yt-dlp==2024.12.23
